REFAC: modified file structure and added folder for easier readability
