+++ src/types/collection.ts
@@ -1,0 +1,7 @@
+export interface Collection {
+  id: string;
+  name: string;
+  description: string;
+  accentColor: string;
+  bookIds: number[];
+}
